Split bump lockfile repository-port tests into their own module
Resolve the CodeScene Low Cohesion finding on
`tests/unit/test_bump_lockfile_rebuild.py` by moving the injected-port
integration tests into a focused module. `_RecordingLockfileRepository`,
`test_run_uses_injected_lockfile_repository`, and
`test_dry_run_projects_through_lockfile_repository` move verbatim into the
new `tests/unit/test_bump_lockfile_repository.py`; the original file keeps
the monkeypatch-based rebuild scenarios and `_LockfileSkipScenario`.

Imports are trimmed to what each file needs: the new module imports only
`cabc`, `pathlib`, `bump`, and the workspace builders, and the original
drops the now-unused `collections.abc` import. No production code changes.
